manual updates from ctl UI
