--- clients/ctl/admin-ui/src/features/taxonomy/index.ts
@@ -1,1 +1,0 @@
-export * from "./taxonomy.slice";
